# Speed up arithmetic kernels, reduce `unsafe` usage