@alamb 2025-05-13
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing arithmetic_speed (df4c5553cebff3ec2b466172f129518434fb4a4e) to 6a3ecefd184b51ac66942ad013395aa882676ca1 [diff](https://github.com/apache/arrow-rs/compare/6a3ecefd184b51ac66942ad013395aa882676ca1..df4c5553cebff3ec2b466172f129518434fb4a4e)
BENCH_NAME=concatenate_kernel
BENCH_COMMAND=cargo bench --all-features --bench concatenate_kernel 
BENCH_FILTER=
BENCH_BRANCH_NAME=arithmetic_speed
Results will be posted here when complete