@alamb 2025-05-13
🤖: Benchmark completed

<details><summary>Details</summary>
<p>


```
group                                        arithmetic_speed                       main
-----                                        ----------------                       ----
concat 1024 arrays boolean 4                 1.00     28.2±0.13µs        ? ?/sec    1.00     28.3±0.05µs        ? ?/sec
concat 1024 arrays i32 4                     1.00     13.2±0.05µs        ? ?/sec    1.00     13.2±0.03µs        ? ?/sec
concat 1024 arrays str 4                     1.00     98.3±0.94µs        ? ?/sec    1.00     97.9±0.90µs        ? ?/sec
concat boolean 1024                          1.01    482.1±1.02ns        ? ?/sec    1.00    475.8±2.33ns        ? ?/sec
concat boolean 8192 over 100 arrays          1.00     50.9±0.10µs        ? ?/sec    1.00     50.9±0.06µs        ? ?/sec
concat boolean nulls 1024                    1.01    848.1±4.49ns        ? ?/sec    1.00    836.1±1.64ns        ? ?/sec
concat boolean nulls 8192 over 100 arrays    1.00    118.7±0.32µs        ? ?/sec    1.00    118.7±0.23µs        ? ?/sec
concat fixed size lists                      1.04   753.2±23.59µs        ? ?/sec    1.00   725.3±33.72µs        ? ?/sec
concat i32 1024                              1.02    470.3±2.12ns        ? ?/sec    1.00    459.9±2.13ns        ? ?/sec
concat i32 8192 over 100 arrays              1.00    209.6±6.73µs        ? ?/sec    1.02   213.3±13.32µs        ? ?/sec
concat i32 nulls 1024                        1.14    989.5±9.18ns        ? ?/sec    1.00   867.2±11.21ns        ? ?/sec
concat i32 nulls 8192 over 100 arrays        1.05    293.2±7.65µs        ? ?/sec    1.00    278.4±3.29µs        ? ?/sec
concat str 1024                              1.00     29.8±0.11µs        ? ?/sec    1.00     29.8±0.23µs        ? ?/sec
concat str 8192 over 100 arrays              1.00    138.1±1.24ms        ? ?/sec    1.00    138.4±1.21ms        ? ?/sec
concat str nulls 1024                        1.01     18.3±0.11µs        ? ?/sec    1.00     18.1±0.04µs        ? ?/sec
concat str nulls 8192 over 100 arrays        1.00     36.0±0.44ms        ? ?/sec    1.99     71.6±0.78ms        ? ?/sec
concat str_dict 1024                         1.00      5.9±0.12µs        ? ?/sec    1.01      6.0±0.05µs        ? ?/sec
concat str_dict_sparse 1024                  1.01      8.3±0.04µs        ? ?/sec    1.00      8.1±0.23µs        ? ?/sec
```


</p>
</details>